pyproject.toml
[project]
name = "control-chart-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "ddgs>=9.5.4",
    "flake8>=7.3.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "matplotlib>=3.10.5",
    "matplotlib-fontja>=1.1.0",
    "networkx>=3.5",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "streamlit>=1.49.0",
    "watchdog>=6.0.0",
]
